requirements.txt
Django==1.11.10
django-crispy-forms==1.7.2
django-debug-toolbar==1.9.1
django-filter==1.1.0
django-like==0.2.0
djangorestframework==3.8.2
Markdown==2.6.11
psycopg2==2.7.4
psycopg2-binary==2.7.4
pytz==2018.4
sqlparse==0.2.4
